frontend: show simple Commentlist also in Overview-Tab
