+++ frontend/src/components/OverviewPage/DiagramTableRow.tsx
@@ -3,6 +3,7 @@ import {BpmnDiagramModel} from "../../model/BpmnDiagramModel";
 import TableCell from "@mui/material/TableCell";
 import TableRow from "@mui/material/TableRow";
 import {Button} from "@mui/material";
+import CommentListSimple from "../CommentListSimple";
 
 export type Props = {
     diagram: BpmnDiagramModel
@@ -20,10 +21,13 @@ export default function DiagramTableRow(props: Props) {
 
     return (
         <>
-            <TableRow key={props.diagram.id} >
+            <TableRow key={props.diagram.id}>
                 <TableCell component="td" scope="row">{props.diagram.name}</TableCell>
                 <TableCell align="left">{props.diagram.filename}</TableCell>
                 <TableCell align="center">{props.diagram.version}</TableCell>
+                <TableCell align="center">
+                    <CommentListSimple diagram={props.diagram}/>
+                </TableCell>
                 <TableCell align="center">{props.diagram.customDiagram ? "yes" : "no"}</TableCell>
                 <TableCell align="center"><Button onClick={handleDetails} color="secondary">Details</Button></TableCell>
             </TableRow>
